perf(menu): native types for analytics schema (date/uuid/smallint buckets)
The view-tracking tables were ported straight from the Bun MVP, which stored
every bucket as text. Tighten them to native Postgres types so keys are ~half
the width and range reads over a day span are index-native:

- Day: text 'YYYY-MM-DD' -> date (DateOnly)      [daily_view, item_view, item_view_seen, menu_session]
- HourBucket: text 'YYYY-MM-DD-HH' -> HourStart timestamptz truncated to the hour [view_seen]
- VisitorId: text (a GUID string) -> uuid         [view_seen, item_view_seen]
- MenuSession.DurationSeconds: int -> smallint    (clamped 1..3600)

Also:
- Drop the redundant SeenAt audit column (+ its index) from both dedup tables —
  the marker's mere existence is the signal; no pruner reads it yet, so the
  index was pure write cost on insert-heavy tables. A retention prune (+ index
  on HourStart / Day) can be added when wanted.
- Restore the counter-table storage tuning the port had dropped (Bun's 0005):
  fillfactor 80 + autovacuum_vacuum_scale_factor 0.02 on daily_view / item_view,
  the UPDATE-heavy tables the public beacon hammers.

Migration rewrites EF's bare ALTER…TYPE into explicit USING casts so real
counter data survives; view_seen (ephemeral dedup markers) is rebuilt fresh.

+++ tests/Iedora.Api.IntegrationTests/MenuTrackingTests.cs
@@ -113,7 +113,7 @@ public sealed class MenuTrackingTests : IntegrationTestBase
         Assert.AreEqual(HttpStatusCode.NoContent, resp.StatusCode);
 
         var duration = await Query(db => db.MenuSessions.Where(s => s.RestaurantId == restId).Select(s => s.DurationSeconds).SingleAsync());
-        Assert.AreEqual(3600, duration); // clamped
+        Assert.AreEqual((short)3600, duration); // clamped
     }
 
     [TestMethod]
